+++ src/main/java/app/komunumo/security/SecurityConfig.java
@@ -84,6 +84,7 @@ public class SecurityConfig extends VaadinWebSecurity {
         return web -> web.ignoring()
                 .requestMatchers(
                         "/.well-known/**",
+                        "/actuator/health",
                         "/images/**");
     }